Completed pro mode, still no styling

+++ server/public/scripts/clientapp.js
@@ -2,12 +2,13 @@ var completion = {};
 $(document).ready(function() {
     addListItems();
     $('#submit').on('click', postItem);
-    $('#todo-items').on('click', '.delete', deleteItem);
-    $('#todo-items').on('click', '.update', updateStatus);
+    $('.todo-items').on('click', '.delete', deleteItem);
+    $('.todo-items').on('click', '.update', updateStatus);
 });
 
 function addListItems() {
-    $('#todo-items').empty();
+    $('#todo-items-complete').empty();
+    $('#todo-items-incomplete').empty();
     $.ajax({
         type: 'GET',
         url: '/todo',
@@ -22,15 +23,20 @@ function addListItems() {
                 });
                 console.log(item.id);
                 $container.data('itemID', item.id);
+                $container.data('name', item.name);
+                $container.data('complete', item.complete);
+                $container.append('<td><button class="delete">Delete</button></td>');
                 if (item.complete) {
                     $container.append('<td>Complete</td><td><button class="unfinished update">Set to Incomplete</button></td>');
                     $container.addClass('complete');
+                    $('#todo-items-complete').append($container);
+
                 } else {
                     $container.append('<td>Incomplete</td><td><button class="finished update">Set to Complete</button></td>');
                     $container.addClass('incomplete');
+                    $('#todo-items-incomplete').append($container);
+
                 }
-                $container.append('<td><button class="delete">Delete</button></td>');
-                $('#todo-items').append($container);
             });
         }
     });
@@ -42,7 +48,6 @@ function postItem() {
     var item = {};
 
     $.each($('#taskform').serializeArray(), function(i, field) {
-        console.log(field.value);
         item[field.name] = field.value;
 
     });
@@ -53,33 +58,33 @@ function postItem() {
         data: item,
         success: function(data) {
             addListItems();
-            console.log('successful post!');
         }
     });
 }
 
 function deleteItem() {
     event.preventDefault();
-    console.log($(this).closest('tr'));
+    var $name = $(this).closest('tr').data('name');
     var itemID = $(this).closest('tr').data('itemID');
+    var delTask = confirm('Are you sure you want to delete ' + $name + '?');
+    if (delTask === true) {
 
-    $.ajax({
-        type: 'DELETE',
-        url: '/todo/' + itemID,
-        success: function() {
-            addListItems();
-        }
-    });
+        $.ajax({
+            type: 'DELETE',
+            url: '/todo/' + itemID,
+            success: function() {
+                addListItems();
+            }
+        });
+    }
 }
 
 function updateStatus() {
     event.preventDefault();
     if ($(this).closest('tr').hasClass('incomplete')) {
         completion.value = true;
-        console.log(completion);
     } else {
         completion.value = false;
-        console.log(completion);
     }
     var itemID = $(this).closest('tr').data('itemID');
     $.ajax({
